fix: Move wallet creation func to sample
