--- app/test_common.go
@@ -1,26 +1,0 @@
-package app
-
-import (
-	"math/rand"
-
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// GeneratePrivKeyAddressPairsFromRand generates (deterministically) a total of n private keys and addresses.
-func GeneratePrivKeyAddressPairs(n int) (keys []cryptotypes.PrivKey, addrs []sdk.AccAddress) {
-	r := rand.New(rand.NewSource(12345)) // make the generation deterministic
-	keys = make([]cryptotypes.PrivKey, n)
-	addrs = make([]sdk.AccAddress, n)
-	for i := 0; i < n; i++ {
-		secret := make([]byte, 32)
-		_, err := r.Read(secret)
-		if err != nil {
-			panic("Could not read randomness")
-		}
-		keys[i] = secp256k1.GenPrivKeyFromSecret(secret)
-		addrs[i] = sdk.AccAddress(keys[i].PubKey().Address())
-	}
-	return
-}
